Changed sizes from uint8 to uint16

+++ ringBuffer.c
@@ -5,7 +5,7 @@ void ring_buffer_init(ring_buffer_t *buffer) {
   buffer->head_index = 0;
 }
 
-void ring_buffer_queue(ring_buffer_t *buffer, uint8_t data) {
+void ring_buffer_queue(ring_buffer_t *buffer, uint16_t data) {
 
   if(ring_buffer_is_full(buffer)) {
     buffer->tail_index = ((buffer->tail_index + 1) & RING_BUFFER_MASK);
@@ -15,7 +15,7 @@ void ring_buffer_queue(ring_buffer_t *buffer, uint8_t data) {
   buffer->head_index = ((buffer->head_index + 1) & RING_BUFFER_MASK);
 }
 
-ring_buffer_size_t ring_buffer_dequeue(ring_buffer_t *buffer, uint8_t *data) {
+ring_buffer_size_t ring_buffer_dequeue(ring_buffer_t *buffer, uint16_t *data) {
   if(ring_buffer_is_empty(buffer)) {
     return 0;
   }
@@ -25,6 +25,6 @@ ring_buffer_size_t ring_buffer_dequeue(ring_buffer_t *buffer, uint8_t *data) {
   return 1;
 }
 
-extern inline uint8_t ring_buffer_is_empty(ring_buffer_t *buffer);
-extern inline uint8_t ring_buffer_is_full(ring_buffer_t *buffer);
-extern inline uint8_t ring_buffer_num_items(ring_buffer_t *buffer);
+extern inline uint16_t ring_buffer_is_empty(ring_buffer_t *buffer);
+extern inline uint16_t ring_buffer_is_full(ring_buffer_t *buffer);
+extern inline uint16_t ring_buffer_num_items(ring_buffer_t *buffer);
